✨ feat:#6 결제페이지 타입 분할 및 .env파일 Portone key값 반영
